Refactor GraphQL schema into modular structure
- Split schema definitions for queries, mutations, and types into separate files for better maintainability.
- Update `GRAPHENE` configuration in settings to reference the new `graphql_schema.py`.
