+++ nbxdjango/packagehandling/graphql_schema.py
@@ -1,0 +1,5 @@
+import graphene
+from .schema.queries import Query
+from .schema.mutations import Mutation
+
+schema = graphene.Schema(query=Query, mutation=Mutation)
